refactor: semantic dom
