+++ docs/component/prompts.md
@@ -100,18 +100,7 @@ type SemanticType = 'list' | 'item' | 'content' | 'title' | 'subList' | 'subItem
 
 ## Semantic DOM
 
-:::semantic
-
-prompts/base
-  - title: 标题容器
-  - list: 列表容器
-  - item: 列表项
-  - itemContent: 列表项内容
-
-prompts/nested
-  - subList: 子列表容器
-  - subItem: 子列表项
-
+<vp-semantic component="Prompts"></vp-semantic>
 
 ## 主题变量（Design Token）
