+++ docs/component/welcome.md
@@ -62,14 +62,6 @@ welcome/background
 
 ## Semantic DOM
 
-:::semantic
-
-welcome/base:
-  - icon: 图标容器
-  - title: 标题容器
-  - description: 描述容器
-  - extra: 额外内容
-
-:::
+<vp-semantic component="Welcome"></vp-semantic>
 
 ## 主题变量（Design Token）
