--- docs/semantics/prompts/base.vue
@@ -1,39 +1,0 @@
-<script setup lang="tsx">
-import { Prompts, type PromptsProps } from 'ant-design-x-vue'
-import {BulbOutlined, InfoCircleOutlined, RocketOutlined} from '@ant-design/icons-vue'
-
-const {classNames} = defineProps<{
-  classNames: PromptsProps['classNames']
-}>()
-
-const items: PromptsProps['items'] = [
-  {
-    key: '1',
-    icon: <BulbOutlined style={{ color: '#FFD700' }} />,
-    label: 'Ignite Your Creativity',
-    description: 'Got any sparks for a new project?',
-    disabled: false,
-  },
-  {
-    key: '2',
-    icon: <InfoCircleOutlined style={{ color: '#1890FF' }} />,
-    label: 'Uncover Background Info',
-    description: 'Help me understand the background of this topic.',
-    disabled: false,
-  },
-  {
-    key: '3',
-    icon: <RocketOutlined style={{ color: '#722ED1' }} />,
-    label: 'Efficiency Boost Battle',
-    description: 'How can I work faster and better?',
-    disabled: false,
-  },
-];
-
-defineRender(() => {
-  return (
-    <Prompts classNames={classNames} title="✨ Inspirational Sparks and Marvelous Tips" items={items} />
-  )
-})
-
-</script>
